implemented load session as callback instead of useEffect
