+++ src/components/LoadSession.tsx
@@ -15,7 +15,7 @@ export default function LoadSessionButton(props: any) {
         reader.onload = (event: any) => {
             try{
                 const jsonObj = JSON.parse(event.target.result);
-                props.setSessionJSON(jsonObj);
+                props.handleLoadFileClick(jsonObj)
             }
             catch(error) {
                 console.error(error)
